% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-metric.R
\name{calc_metric}
\alias{calc_metric}
\title{Calculate Macro-invertebrate Metrics}
\usage{
calc_metric(
  data,
  metrics = c("awic", "epsi", "psi", "riverfly", "spear", "whpt"),
  taxa_list = "TL2",
  taxon_table = macroinvertebrateMetrics::macroinvertebrateTaxa,
  names = macroinvertebrateMetrics::column_attributes$name,
  questions = c("Taxon abundance", "Taxon Abundance", "Live abundance"),
  metric_cols = macroinvertebrateMetrics::metric_cols,
  ...
)
}
\arguments{
\item{data}{dataframe like `demo_data`}

\item{metrics}{One One or more of "awic", "epsi", "riverfly", "spear", "whpt".}

\item{taxa_list}{Taxa list either "TL2", "TL3" or "TL5.}

\item{taxon_table}{Optional Dataframe with WHPT scores and taxa. Default is NULL and will use
built in WHPT scores. But you could supply custom dataframe if required for
experimenation/development purposes.}

\item{questions}{Optional, user provided 'question' default is 'Taxon abundance', which
filters only abundance values.}

\item{metric_cols}{Columns used from taxon table to calculate metrics plus to
taxon name column to join to input data (if using custom taxon table)}

\item{...}{Pass in specific paramters for each metric.}
}
\value{
dataframe
}
\description{
Calculate Macro-invertebrate Metrics
}
\examples{
output <- calc_metric(demo_data)
}
