% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-psi.R
\name{calc_psi}
\alias{calc_psi}
\title{Proportion of Sediment-sensitive Invertebrates (PSI)}
\usage{
calc_psi(ecologyResults, taxaList = "TL3")
}
\arguments{
\item{ecologyResults}{Dataframe with at least three columns
item{SAMPLE_ID} - unique idenftier for each sample
item{TAXON} - Taxon name that matches to macroinvertebrateTaxa dataset
item{RESULT} - Log abundance category}

\item{taxaList}{The taxonomic level the sample(s) have been identified at
according to specificed taxa lists as described in WFD100 Further
Development of River Invertebrate Classification Tool. Either "TL3" - Taxa
list 3, "TL4" - Taxa list 4 or  "TL5" - Taxa list 5.}
}
\value{
Dataframe with
item{SAMPLE_ID}
item{ANALYSIS_NAME}
item{DETERMINAND}
item{RESULT}
}
\description{
A sediment-sensitive macro-invertebrate metric that provides a proxy to
describe the extent to which the surface of river bed are composed, or
covered by sediments. It can be calculated at Taxonomic Levels 3, 4 & 5
}
\examples{
sample <- demoEcologyResults
sample <- filter_psi(sample, taxaList = "TL3")
calc_psi(ecologyResults = sample)
}
\references{
Extence, Chris & Chadd, Richard & England, Judy & Dunbar, M.J. & Wood, Paul &
Taylor, E.D.. (2010). The Assessment of Fine Sediment Accumulation in Rivers
Using Macro-Invertebrate Community Response. River Research and Applications.
29. 10.1002/rra.1569.
}
