% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-whpt.R
\name{calc_whpt}
\alias{calc_whpt}
\title{Calculate WHPT scores}
\usage{
calc_whpt(ecologyResults, taxonTable = NULL)
}
\arguments{
\item{ecologyResults}{Dataframe of taxonomic results with mandatory three columns:
\describe{
  \item{SAMPLE_ID}{Unique sample identifier}
  \item{RESULT}{Numeric abundance}
  \item{TAXON}{Character - TL2 WHPT taxon name}
  }
Columns names must match these exactly, but the column order does not matter.
See demoEcologyResults for example dataset.}

\item{taxonTable}{Optional Dataframe with WHPT scores and taxa. Default is NULL and will use
built in WHPT scores. But you could supply custom dataframe if required for
experimenation/development purposes.}
}
\value{
Dataframe with four columns: SAMPLE_ID, DETERMINAND, RESULT, ANALYSIS_NAME,
ANALYSIS_REPNAME
}
\description{
Calculate WHPT scores
}
\examples{
metricResults <- calc_whpt(demoEcologyResults)
}
