% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASI.R
\name{CalcASI}
\alias{CalcASI}
\title{Riverfly Angler's Score Index}
\usage{
CalcASI(data, season = NULL, TL = 3L)
}
\arguments{
\item{data}{A dataframe containing \emph{standardised} taxa with at least
eight columns and in the specified order. Extra columns can be added,
but they must be related to Sampling Point. See 'Details'}

\item{season}{Optional. An integer vector containing up to 7 elements,
from \samp{1L} to \samp{7L}. It is only needed when the input data frame contains
Season and Year instead of Sample ID and Date. See 'Details' for more info about
input columns.}

\item{TL}{An integer vector containing code \samp{3}. Which is the closest TL
to include the Angler's Score Index in. \samp{3} is the default value anyway,
because this index actually is not defined in any specific TL, so it can be ignored.}
}
\value{
A dataframe with the following columns: \code{TL}, \code{Site},
\code{Season} or \code{Sample ID}, \code{Year} or \code{Date} (Date object),
\code{Infered}, extra columns - if any, and \code{ASI} (Angler's Score Index).
}
\description{
The Riverfly Partnership monitoring initiative aims to address perceived
declines in the abundance of target benthic macroinvertebrates (mostly river
flies).
}
\details{
The input dataframe \code{data} must contain a minimum of eight columns in the
specified order: \code{TL} (Taxonomic Level), \code{Site}, \code{Season} or \code{Sample ID},
\code{Year} or \code{Date} (Date object), \code{Maitland Code}, \code{Maitland Name}, \code{Abundance},
\code{Infered} (Code returned by the function \code{StandardiseRawTaxa}
which can be declared as \samp{0} if \code{StandardiseRawTaxa} is not used).

If the data frame contains Season and Year instead of Sample ID / Date, the \code{season}
parameter must be a vector of the form \samp{c(1L, 3L, 5L)}, indicating which seasons, and
combination of seasons, must be computed.

Although there is no output TL defined for this index, a default value of \samp{3}
has been chosen to keep the same signature as similar functions in the package.

The available codes for \code{Season} and a general description of Taxonomic
Levels are defined in the main help page of the \code{aquaMetrics} package.
}
\note{
The extra columns can be used to aggregate data by different criteria, i.e.,
other than season or Sample ID. For doing so you must fill all season fields with a dummy
value \samp{1}. Thus, the aggregation will disregard season/Sample ID as an aggregation column.
}
\examples{
\dontrun{
#No examples yet
}
}
\seealso{
\code{\link{StandardiseRawTaxa}}
}
