% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC.R
\name{CalcCC}
\alias{CalcCC}
\title{Climate Change Metric}
\usage{
CalcCC(data, season = NULL, TL = 5L,
       method = 1L)
}
\arguments{
\item{data}{A dataframe containing \emph{standardised} taxa with at least
eight columns and in the specified order. Extra columns can be added,
but they must be related to Sampling Point. See 'Details'}

\item{season}{Optional. An integer vector containing up to 7 elements,
from \samp{1L} to \samp{7L}. It is only needed when the input data frame contains
Season and Year instead of Sample ID and Date. See 'Details' for more info about
input columns.}

\item{TL}{An integer vector containing up to 2 elements which would
represent the taxonomic level to calulate Climate Change at.}

\item{method}{It can take either take the numeric value \samp{1L}, which
computes the percentage of indicators, or \samp{2} which calculates the
proportion of \emph{n} thermal indicators (useful when data contain log
abundances rather than counts).}
}
\value{
A dataframe with the following columns: \code{TL}, \code{Site},
\code{Season} or \code{Sample ID}, \code{Year} or \code{Date} (Date object),
\code{Infered}, extra columns - if any, and \emph{CC metrics} which contain
\code{CS} (Cold Stenotherms), \code{WS} (Warm Stenotherms), \code{E} (Eurytherms) and
\code{TIndicators} (Percentage or proportion of thermal indicators in the
sample).
}
\description{
A temperature metric that computes the ratio or porportion of organisms with
different temperature tolerances. It can be calculated at Taxonomic Levels 5 and 4
}
\details{
The input dataframe \code{data} must contain a minimum of eight columns in the
specified order: \code{TL} (Taxonomic Level), \code{Site}, \code{Season} or \code{Sample ID},
\code{Year} or \code{Date} (Date object), \code{Maitland Code}, \code{Maitland Name}, \code{Abundance},
\code{Infered} (Code returned by the function \code{StandardiseRawTaxa}
which can be declared as \samp{0} if \code{StandardiseRawTaxa} is not used).
If the data frame contains Season and Year instead of Sample ID / Date, the \code{season}
parameter must be a vector of the form \samp{c(1L, 3L, 5L)}, indicating which seasons, and
combination of seasons, must be computed.
This function works with both current Maitland Codes and new ones,
defined as of JUL-2014.

The available codes for \code{Season} and a general description of Taxonomic
Levels are defined in the main help page of the \code{aquaMetrics} package.
}
\note{
The extra columns can be used to aggregate data by different criteria, i.e.,
other than season or Sample ID. For doing so you must fill all season fields with a dummy
value \samp{1}. Thus, the aggregation will disregard season/SampleID as an aggregation column.
}
\examples{
\dontrun{
#No examples yet
}
}
\seealso{
\code{\link{StandardiseRawTaxa}}
}
