% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEpsi.R
\name{calcEpsi}
\alias{calcEpsi}
\title{Enchanced Proportion of Sediment-sensitive Invertebrates (EPSI)}
\usage{
calcEpsi(ecologyResults, taxaList = "TL2", logAbundance = TRUE)
}
\arguments{
\item{ecologyResults}{Dataframe with at least two columns
item{SAMPLE_ID} - unique idenftier for each sample
item{TAXON} - Taxon name that matches to macroinvertebrateTaxa dataset}
}
\value{
Dataframe with
item{SAMPLE_ID}
item{ANALYSIS_NAME}
item{DETERMINAND}
item{RESULT}
}
\description{
A sediment-sensitive macro-invertebrate metric that provides a proxy to
describe the extent to which the surface of river bed are composed, or
covered by sediments. It can be calculated at Taxonomic Levels 2 & 5
}
\examples{
sample <- demoEcologyResults
sample <- filterPsi(sample,taxaList = "TL3")
calcPsi(ecologyResults= sample)
}
