% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIFE.R
\name{CalcLIFE}
\alias{CalcLIFE}
\title{Lotic-invertebrate Index for Flow Evaluation}
\usage{
CalcLIFE(data, season = NULL, TL = 2L)
}
\arguments{
\item{data}{A dataframe containing \emph{standardised} taxa with at least
eight columns and in the specified order. Extra columns can be added,
but they must be related to Sampling Point. See 'Details'}

\item{season}{Optional. An integer vector containing up to 7 elements,
from \samp{1L} to \samp{7L}. It is only needed when the input data frame contains
Season and Year instead of Sample ID and Date. See 'Details' for more info about
input columns.}

\item{TL}{An integer vector containing up to 3 elements which would
represent the taxonomic level to calulate LIFE at. They can be \samp{1L}
(for TL1/2), \samp{2L} and \samp{5L}.}
}
\value{
A dataframe with the following columns: \code{TL}, \code{Site},
\code{Season} or \code{Sample ID}, \code{Year} or \code{Date} (Date object),
\code{Infered}, extra columns - if any, and \code{LIFEScore}.
}
\description{
The LIFE method is based on recognized flow associations of different
macroinvertebrate taxa. It links qualitative and semi-quantitative change
in riverine benthic macroinvertebrate communities to prevailing flow regimes.
It can be calculated at Taxonomic Levels 5, 2 and 1 (for TL1/2 composite
families).
}
\details{
The input dataframe \code{data} must contain a minimum of eight columns in the
specified order: \code{TL} (Taxonomic Level), \code{Site}, \code{Season} or \code{Sample ID},
\code{Year}  or \code{Date} (Date object), \code{Maitland Code}, \code{Maitland Name},
\code{Abundance}, \code{Infered} (Code returned by the function \code{StandardiseRawTaxa}
which can be declared as \samp{0} if \code{StandardiseRawTaxa} is not used).

If the data frame contains Season and Year instead of Sample ID / Date, the \code{season}
parameter must be a vector of the form \samp{c(1L, 3L, 5L)}, indicating which seasons, and
combination of seasons, must be computed.

The available codes for \code{Season} and a general description of Taxonomic
Levels are defined in the main help page of the \code{aquaMetrics} package.
}
\note{
More extra columns can be added to aggregate data by different criteria, i.e.,
other than season or Sample ID. For doing so you must fill all season fields with a dummy
value: \samp{1}.Thus, the aggregation would disregard season/Sample ID as an aggregation column.
}
\examples{
\dontrun{
#No examples yet
}
}
\references{
Extence, C, Balbi, D., Chadd, R. 1999. \emph{River flow indexing using
British benthic macroinvertebrates: A framework for setting hydroecological
objectives}. Regul Rivers Res Mgmt. 15: 543-574.
}
\seealso{
\code{\link{StandardiseRawTaxa}}
}
