% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPsi.R
\name{calcPsi}
\alias{calcPsi}
\title{Calculate Pollution Silt Index (PSI)}
\usage{
calcPsi(ecologyResults)
}
\arguments{
\item{ecologyResults}{Dataframe of Ecology results}
}
\value{
dataframe with...
}
\description{
Calculate Pollution Silt Index (PSI)
}
\examples{
calcEpsi(demoEcologyResults)
}
