% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiverfly.R
\name{calcRiverfly}
\alias{calcRiverfly}
\title{Riverfly metric}
\usage{
calcRiverfly(ecologyResults = ecologyResults)
}
\arguments{
\item{ecologyResults}{A data frame of ecology view data}
}
\value{
A data frame 5 variables
}
\description{
Function to calculate riverfly scores
}
\examples{
ecologyResults <- demoEcologyResults
ecologyResults <- ecologyResults[ecologyResults$ANALYSIS_NAME == "FW_TAX_ID",]
spearOutput <- calcRiverfly(ecologyResults)
}
