% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSpear.R
\name{calcSpear}
\alias{calcSpear}
\title{Calculate SPEAR metric}
\usage{
calcSpear(ecologyResults)
}
\arguments{
\item{ecologyresults}{Dataframe of results with four columns: SAMPLE_NUMBER, TAXON, SPEAR_SPECIES, RESULT}

\item{recoveryArea}{There are 3 different paramaters depending on availability of "Recovery areas" information:
\itemize{
  \item "Absent" - No presence of recovery areas
  \item "Present" - Presence of recovery areas
  \item "Unknown" - No information available
}}
}
\value{
dataframe with...
}
\description{
Calculate SPEAR metric
}
\examples{
calcSpear(demoEcoloyResults, recoveryArea = "unknown")
}
