% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWhpt.R
\name{calcWhpt}
\alias{calcWhpt}
\title{Calculate WHPT scores}
\usage{
calcWhpt(ecologyResults, taxonTable = NULL)
}
\arguments{
\item{ecologyResults}{dataframe of taxonomic results with mandatory four columns
SAMPLE_ID, DETERMINAND, RESULT, TAXON}

\item{taxonTable}{dataframe with WHPT scores and taxa}
}
\value{
Dataframe with four columns: SAMPLE_ID, DETERMINAND, RESULT, ANALYSIS_NAME, ANALYSIS_REPNAME
}
\description{
Calculate WHPT scores
}
\examples{
metricResults <- calcWhpt(demoEcologyResults)
}
