% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo Ecology results for testing purposes}
\format{
A data frame
}
\usage{
demo_data
}
\description{
Dataframe containing ecology results including macroinvertebrate data for
testing and examples.
}
\keyword{datasets}
