% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoEcologyResults.R
\docType{data}
\name{demoEcologyResults}
\alias{demoEcologyResults}
\title{Demo Ecology results for testing purposes}
\format{
A data frame
}
\usage{
demoEcologyResults
}
\description{
Dataframe containing ecology results including macroinvertebrate data for
testing and examples.
}
\keyword{datasets}
