% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertebrateMetrics.R
\docType{package}
\name{invertebrateMetrics}
\alias{invertebrateMetrics}
\alias{invertebrateMetrics-package}
\title{Freshwater Biotic indices}
\description{
Metrics for assessing ecological indicators of biodiversity in rivers and lakes.
Specifically coded to be used with TIBCO Enterprise Runtime for R (TERR).
}
\details{
\tabular{ll}{
 Package: \tab invertebrateMetrics\cr
 Type: \tab Package\cr
 Version: \tab 0.98-2\cr
 Date: \tab 2017-03-30\cr
 License: \tab MIT\cr
}

This package contains several functions which compute multiple ecological
metrics at different Taxonomic Levels (TL) using multiple season combinations.
\itemize{
 \item SPEAR (\code{\link{calcSpear}})\cr Species at Risk
 \item PSI (\code{\link{calcPsi}})\cr Proportion of Sediment-sensitive Invertebrates
 \item LIFE (\code{\link{CalcLIFE}})\cr Lotic Invertebrate Index for Flow Evaluation
 \item Riverfly's ASI (\code{\link{CalcASI}})\cr Riverfly Angler's Score Index
 \item Climate Change (\code{\link{CalcCC}})\cr Climate Change. Thermal indicator.
}

Taxonomic Levels are the different output options developed by RIVAPCS
to produce results at both species and family levels:\cr
TL1 - BMWP families\cr
TL2 - Revised BMWP (WHPT) families\cr
TL3 - All families\cr
TL4 - RIVPACS species\cr
TL5 - WFD species

Season codes define different season combinations used to calculate the
indices, those are listed below:\cr
\samp{1} = spring\cr
\samp{2} = summer\cr
\samp{3} = autumn\cr
\samp{4} = spring + summer\cr
\samp{5} = spring + autumn\cr
\samp{6} = summer + autumn\cr
\samp{7} = spring + summer + autumn

There is also a common routine (\code{\link{StandardiseRawTaxa}}) which should
be run prior to any calculation in order to standardise taxa counts into the
required Taxonomic Levels.
}
\author{
Scottish Environment Protection Agency

Maintainer: Carlos Ruiz \email{carlos.ruiz@sepa.org.uk}
}
\keyword{package}
