% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertebrateMetrics.R
\docType{data}
\name{LIFEFlowScores}
\alias{LIFEFlowScores}
\title{LIFE. Scores of taxa associated with flow groups}
\format{A data frame with 36 rows and 4 variables}
\description{
Scores for different abundance categories of taxa associated with flow groups 1-6
}
\details{
The columns are as follows:

\itemize{
 \item Flow.Group.
 \item Flow.Group.Description.
 \item Log10.Abundance.Category. Based on log10(abundance).
 \item fs.  Taxon flow score.
}
}
\keyword{datasets}
