% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertebrateMetrics.R
\docType{data}
\name{LookUpASI}
\alias{LookUpASI}
\title{Lookup ASI}
\format{A data frame with 33 rows and 4 variables}
\description{
Lookup table of taxa classification for Riverfly's ASI analysis
}
\details{
The columns are as follows:

\itemize{
 \item TL. Taxonomic level of the taxon.
 \item CODE. Maitland Code.
 \item NAME.  Name of the taxon.
 \item ASI.  Target group
}
}
\keyword{datasets}
