% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertebrateMetrics.R
\docType{data}
\name{LookUpCC}
\alias{LookUpCC}
\title{Lookup Climate Change}
\format{A data frame with 246 rows and 5 variables}
\description{
Lookup table of taxa classification for Climate Change analysis
}
\details{
The columns are as follows:

\itemize{
 \item TL. Taxonomic level of the taxon.
 \item CODE. Maitland Code.
 \item NAME.  Name of the taxon.
 \item TEMP.  Temperature tolerance
 \item NEW.CODE.  New Maitland Code
}
}
\keyword{datasets}
