% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroinvertebrateTaxa.R
\docType{data}
\name{macroinvertebrateTaxa}
\alias{macroinvertebrateTaxa}
\title{Macroinvertebrate Taxa}
\format{A data frame}
\usage{
macroinvertebrateTaxa
}
\description{
Dataframe of all macroinvertebrate taxa used form WFD100 / RICT
}
\keyword{datasets}
