% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_cols.R
\docType{data}
\name{metric_cols}
\alias{metric_cols}
\title{Metrics Metadata}
\format{
A data frame
}
\usage{
metric_cols
}
\value{
A data frame with 2 columns and 18 rows, containing
this key columns:
\itemize{ \item
  \item name - metric name
  \item metric_names  - column(s) containing metric score
}
}
\description{
Dataframe containing metadata about the metrics
}
\keyword{datasets}
