% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiCondition.R
\docType{data}
\name{psiCondition}
\alias{psiCondition}
\title{PSI condition definition table}
\format{
A data frame
}
\usage{
psiCondition
}
\description{
Dataframe containing the definition of the PSI conditions
}
\keyword{datasets}
