% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertebrateMetrics.R
\docType{data}
\name{PSISensitivity}
\alias{PSISensitivity}
\title{PSI. Fine Sediment Sensitivity Rating definitons and abundance}
\format{A data frame with 30 rows and 4 variables}
\description{
Fine Sediment Sensitivity Rating definitons and abundance weighted scores for
PSI calculation
}
\details{
The columns are as follows:

\itemize{
 \item GROUP. Sediment-sensitivity group.
 \item DESCRIPTION.
 \item LOG10.ABUN.CAT.  log10(abundance) category
 \item SEDIMENT.SS.  Abundance weighted scores
}
}
\keyword{datasets}
