% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiSensitivityScore.R
\docType{data}
\name{psiSensitivityScore}
\alias{psiSensitivityScore}
\title{PSI Sensitivity Score table}
\format{
A data frame
}
\usage{
psiSensitivityScore
}
\description{
Dataframe containing the definition of the PSI Sensitivity Scores
}
\keyword{datasets}
