% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverfly.R
\name{riverfly}
\alias{riverfly}
\title{Riverfly metric}
\usage{
riverfly(
  data,
  names = macroinvertebrateMetrics::column_attributes$name,
  questions = c("Taxon abundance", "Taxon Abundance", "Live abundance"),
  metric_cols = metric_cols
)
}
\arguments{
\item{data}{A data frame of ecology data as defined in `column_attributes`.}

\item{names}{Optional, user provided list of column names different to those used
in `column_attributes` to match with input data}

\item{questions}{Optional, user provided 'question' default is 'Taxon abundance', which
filters only abundance values.}
}
\value{
A data frame 5 variables
}
\description{
Function to calculate riverfly scores
}
\examples{
data <- demo_data
data <- data[data$parameter == "River Family Inverts", ]
output <- calc_riverfly(data)

}
