% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardiseRawTaxa.R
\name{StandardiseRawTaxa}
\alias{StandardiseRawTaxa}
\title{Standardise Raw Taxa}
\usage{
StandardiseRawTaxa(family.df = NULL, species.df = NULL,
                   infer.lower.levels = FALSE,
                   aggregate = "sample")
}
\arguments{
\item{family.df}{An optional data frame containing family-level data from
NEMS (Analysis: FW_TAX_ID) with at least six columns and in the specified
order. Extra columns can be added, but they must be only related to
Sampling Point. See 'Details'.}

\item{species.df}{An optional data frame containing species-level data from
NEMS (Analysis: MIXTAX_TST) with at least six columns and in the specified
order. Extra columns can be added, but they must be only related to
Sampling Point. See 'Details'.}

\item{infer.lower.levels}{Flag (\samp{TRUE}/\samp{FALSE}) to infer Taxonomic
Levels at TL1, TL2 and TL3 from species-level data.}

\item{aggregate}{Indicates whether the aggregation is either done by \code{"sample"}
or by \code{"season"}. Depending on the aggregation, input data frames should be
composed of different type of columns.}
}
\value{
A list containing 3 data frames.

 \code{standard.taxa}: Dataframe containing the following columns (depending on input data):
   \code{TL} (integer), \code{Sampling Point}, \code{Season} or \code{Sample},
   \code{Year} or \code{Date}, \code{Maitland Code},
   \code{Name}, \code{Abundance}, \code{Infered} (integer).\cr
   The column \code{Infered} indicates whether TL1/TL2/TL3 were infered from
   data in \code{species.df} If the user entered extra columns within the
   input dataframes they will be included here as well.

 \code{not.included.families}: Dataframe (which can be empty) containing those
   input families that have not been found in any of the TL lists.

 \code{not.included.species}: Dataframe (which can be empty) containing those
   input species that have not been found in any of the TL lists.
}
\description{
This function standardise taxa retrieved from NEMS to all available Taxon Levels.
Most of the times this function will be run to generate the input data required
by the other functions present in this package.
}
\details{
Both \code{family.df} and \code{species.df} are data frames containing at least
the following six columns, depending on the 'aggregate' parameter:\cr
\samp{"season"}: \code{Site}, \code{Season} (integer), \code{Year},
\code{Maitland Code}, \code{Maitland Name}, \code{Abundance}.\cr
\samp{"sample"}: \code{Site}, \code{Sample Number}, \code{Sample Date} (Date object),
\code{Maitland Code}, \code{Maitland Name}, \code{Abundance}.\cr
Actually, the routine performs the same calculations for both types of aggregation,
the only change is the name of the output data frame's columns.

The available codes for \code{Season} and a general description of Taxonomic
Levels are defined in the main help page of the \code{macroinvertebrateMetrics} package.
}
\note{
More extra columns can be added to aggregate data by different criteria, i.e.,
other than season/date. For doing so you must fill all season/date fields with a dummy value: 1.
Thus, the aggregation would disregard season/date as an aggregation column.
}
\examples{
\dontrun{
#No examples yet
}
}
