% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertebrateMetrics.R
\docType{data}
\name{TaxaList}
\alias{TaxaList}
\title{Taxa lists from NEMS}
\format{A data frame with 6491 rows and 18 variables}
\description{
Lookup table of taxa found in NEMS at both family and species level.
}
\details{
The columns are as follows:

\itemize{
 \item Taxon name.
 \item TAXON GROUP.
 \item FAMILY.
 \item MAITLAND CODE.
 \item ARTIFICIAL TAXON.
 \item Metrics Scores (LIFE GROUP, LIFE COMPOSITE, PSI GROUP, SPEAR SPECIES).
 \item Taxons (TL5 TAXON, TL3 FAMILY, TL2 FAMILY, TL1 FAMILY).
 \item MAITLAND_CODE (including leading 0, up to 8 digits).
 \item TL codes (TL5 CODE, TL3 CODE, TL2 CODE, TL1 CODE)
}
}
\keyword{datasets}
