% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whpt.R
\name{whpt}
\alias{whpt}
\title{Calculate WHPT scores}
\usage{
whpt(
  data,
  taxon_table = macroinvertebrateMetrics::macroinvertebrateTaxa,
  names = macroinvertebrateMetrics::column_attributes$name,
  questions = c("Taxon abundance", "Taxon Abundance", "Live abundance"),
  metric_cols = macroinvertebrateMetrics::metric_cols
)
}
\arguments{
\item{taxon_table}{Optional Dataframe with WHPT scores and taxa. Default is NULL and will use
built in WHPT scores. But you could supply custom dataframe if required for
experimenation/development purposes.}

\item{names}{Optional, user provided list of column names different to those used
in `column_attributes` to match with input data}

\item{questions}{Optional, user provided 'question' default is 'Taxon abundance', which
filters only abundance values.}

\item{ecologyResults}{Dataframe of taxonomic results with mandatory three columns:
\describe{
  \item{SAMPLE_ID}{Unique sample identifier}
  \item{RESULT}{Numeric abundance}
  \item{TAXON}{Character - TL2 WHPT taxon name}
  }
Columns names must match these exactly, but the column order does not matter.
See demoEcologyResults for example dataset.}
}
\value{
Dataframe with four columns: SAMPLE_ID, DETERMINAND, RESULT, ANALYSIS_NAME,
ANALYSIS_REPNAME
}
\description{
Calculate WHPT scores
}
\examples{
metricResults <- calc_whpt(demo_data)
}
