% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-awic.R
\name{calc_awic}
\alias{calc_awic}
\title{Acid WFD-AWIC metric}
\usage{
calc_awic(
  data,
  names = macroinvertebrateMetrics::column_attributes$name,
  questions = c("Taxon abundance", "Taxon Abundance"),
  metric_cols = macroinvertebrateMetrics::metric_cols
)
}
\arguments{
\item{data}{dataframe containing mixtaxon invertebrates}

\item{names}{Optional, user provided list of column names different to those used
in `column_attributes` to match with input data}

\item{questions}{Optional, user provided 'question' default is 'Taxon abundance', which
filters only abundance values.}
}
\value{
dataframe
}
\description{
Acid WFD-AWIC metric
}
\examples{
metricResults <- calc_awic(demo_data)
}
