% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcawic.R
\name{calc_awic}
\alias{calc_awic}
\title{Acid WFD-AWIC metric}
\usage{
calc_awic(data)
}
\arguments{
\item{data}{dataframe containing mixtaxon invertebrates}
}
\value{
dataframe
}
\description{
Acid WFD-AWIC metric
}
\examples{
metricResults <- calc_awic(demoEcologyResults)
}
