% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-metric.R
\name{calc_metric}
\alias{calc_metric}
\title{Calculate Macro-invertebrate Metrics}
\usage{
calc_metric(
  data,
  metrics = c("awic", "epsi", "riverfly", "spear", "whpt"),
  taxa_list = "TL2"
)
}
\arguments{
\item{data}{dataframe like `demoEcologyResults`}

\item{metrics}{One One or more of "awic", "epsi", "riverfly", "spear", "whpt".}

\item{taxa_list}{Taxa list either "TL2", "TL3" or "TL5.}
}
\value{
dataframe
}
\description{
Calculate Macro-invertebrate Metrics
}
\examples{
output <- calc_metric(demoEcologyResults)
}
