% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-riverfly.R
\name{calc_riverfly}
\alias{calc_riverfly}
\title{Riverfly metric}
\usage{
calc_riverfly(ecologyResults)
}
\arguments{
\item{ecologyResults}{A data frame of ecology data}
}
\value{
A data frame 5 variables
}
\description{
Function to calculate riverfly scores
}
\examples{
ecologyResults <- demoEcologyResults
ecologyResults <- ecologyResults[ecologyResults$ANALYSIS_NAME == "FW_TAX_ID", ]
spearOutput <- calc_riverfly(ecologyResults)

}
