% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPsi.R
\name{calcPsi}
\alias{calcPsi}
\title{Proportion of Sediment-sensitive Invertebrates (PSI)}
\usage{
calcPsi(ecologyResults, valueWarning = FALSE, taxaList = "TL2")
}
\arguments{
\item{ecologyResults}{Dataframe with at least two columns
item{SAMPLE_ID} - unique idenftier for each sample
item{TAXON} - Taxon name that matches to macroinvertebrateTaxa dataset}
}
\value{
Dataframe with
item{SAMPLE_ID}
item{ANALYSIS_NAME}
item{DETERMINAND}
item{RESULT}
}
\description{
A sediment-sensitive macro-invertebrate metric that provides a proxy to
describe the extent to which the surface of river bed are composed, or
covered by sediments. It can be calculated at Taxonomic Levels 3, 4 & 5
}
\examples{
sample <- demoEcologyResults
sample <- filterPsi(sample,taxaList = "TL2")
calcPsi(ecologyResults= sample)
}
