% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSI.R
\name{CalcPSI}
\alias{CalcPSI}
\title{Proportion of Sediment-sensitive Invertebrates}
\usage{
CalcPSI(data, season = NULL, TL = 3L)
}
\arguments{
\item{data}{A dataframe containing \emph{standardised} taxa with at least
eight columns and in the specified order. Extra columns can be added,
but they must be related to Sampling Point. See 'Details'}

\item{season}{Optional. An integer vector containing up to 7 elements,
from \samp{1L} to \samp{7L}. It is only needed when the input data frame contains
Season and Year instead of Sample ID and Date. See 'Details' for more info about
input columns.}

\item{TL}{An integer vector containing up to 3 elements which would
represent the taxonomic level to calulate PSI at.}
}
\value{
A dataframe with the following columns: \code{TL}, \code{Site},
\code{Season} or \code{Sample ID}, \code{Year} or \code{Date} (Date object),
\code{Infered}, extra columns - if any, \code{PSIScore} and \code{R.Bed.Condition}.
}
\description{
A sediment-sensitive macro-invertebrate metric that provides a proxy to
describe the extent to which the surface of river bed are composed, or
covered by sediments. It can be calculated at Taxonomic Levels 3, 4 & 5.
}
\details{
The input dataframe \code{data} must contain a minimum of eight columns in the
specified order: \code{TL} (Taxonomic Level), \code{Site}, \code{Season} or \code{Sample ID},
\code{Year} or \code{Date} (Date object), \code{Maitland Code}, \code{Maitland Name}, \code{Abundance},
\code{Infered} (Code returned by the function \code{StandardiseRawTaxa}
which can be declared as \samp{0} if \code{StandardiseRawTaxa} is not used).

If the data frame contains Season and Year instead of Sample ID / Date, the \code{season}
parameter must be a vector of the form \samp{c(1L, 3L, 5L)}, indicating which seasons, and
combination of seasons, must be computed.

The available codes for \code{Season} and a general description of Taxonomic
Levels are defined in the main help page of the \code{aquaMetrics} package.
}
\note{
More extra columns can be added to aggregate data by different criteria, i.e.,
other than season or Sample ID. For doing so you must fill all season fields with a dummy
value: \samp{1}.Thus, the aggregation would disregard season/Sample ID as an aggregation column.

Please note that this application applies some changes to standardised families
included in \emph{Ancylidae} and \emph{Oligochaeta} to accomodate the results
to those displayed in RIVPACS database (even though it may not be technically correct).
}
\examples{
\dontrun{
#No examples yet
}
}
\references{
Extence C., Chadd R., England J., Dunbar M., Wood P., Taylor, E. 2013.
\emph{The assessment of fine sediment accumulation in rivers using macro-invertebrate
community response}. River Res Applic. 29: 17-55.
}
\seealso{
\code{\link{StandardiseRawTaxa}}
}
