% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPEAR.R
\name{CalcSPEAR}
\alias{CalcSPEAR}
\title{SPEcies At Risk}
\usage{
CalcSPEAR(data, season = NULL, TL = 2L,
          recovery.area.info = FALSE)
}
\arguments{
\item{data}{A dataframe containing \emph{standardised} taxa with at least
eight columns and in the specified order. Extra columns can be added,
but they must be related to Sampling Point. See 'Details'}

\item{season}{Optional. An integer vector containing up to 7 elements,
from \samp{1L} to \samp{7L}. It is only needed when the input data frame contains
Season and Year instead of Sample ID and Date. See 'Details' for more info about
input columns.}

\item{TL}{An integer vector containing up to 3 elements which would
represent the taxonomic level to calulate SPEAR at.}

\item{recovery.area.info}{Flag (\code{TRUE} / \code{FALSE}) to inform the
program whether the recovery area information has been added to the input parameters.}
}
\value{
A dataframe with the following columns: \code{TL}, \code{Site},
\code{Season} or \code{Sample ID}, \code{Year} or \code{Date} (Date object),
\code{Infered}, \code{Recovery Area information}, \code{SPEAR index},
\code{Water quality classes}, \code{Calculated Toxic Exposure}
(see 'Note'), \code{Aggregated SPEAR abundance} and \code{Aggregated SPEAR
Log(Abundance)}. The last two columns are merely informative and not required
when showing SPEAR results.
}
\description{
Indicator based on biological traits used to detect effects of pesticides on
non-target freshwater invertebrate organisms. It can be calculated at Taxonomic
Levels 2, 4 and 5.
}
\details{
The input dataframe \code{data} must contain a minimum of eight columns in the
specified order: \code{TL} (Taxonomic Level), \code{Site}, \code{Season} or \code{Sample ID},
\code{Year} or \code{Date} (Date object), \code{Maitland Code}, \code{Maitland Name}, \code{Abundance},
\code{Infered} (Code returned by the function \code{StandardiseRawTaxa}
which can be declared as \samp{0} if \code{StandardiseRawTaxa} is not used).
The presence of Recovery area Information can be included adding a ninth column
named \code{Rec} whose possible values are \code{NULL}, \code{1} or \code{0},
for \emph{unknown}, \emph{presence} and \emph{no presence}, respectively.

If the data frame contains Season and Year instead of Sample ID / Date, the \code{season}
parameter must be a vector of the form \samp{c(1L, 3L, 5L)}, indicating which seasons, and
combination of seasons, must be computed.

The available codes for \code{Season} and a general description of Taxonomic
Levels are defined in the main help page of the \code{aquaMetrics} package.
}
\note{
More extra columns can be added to aggregate data by different criteria, i.e.,
other than season or Sample ID. For doing so you must fill those fields with a dummy value:
 \samp{1}. Thus, the aggregation would disregard season/Sample ID as an aggregation column.

The computed \code{Calculated Toxicant Exposure} is only valid for sampling times
during and shortly after the main agricultural insecticide use (first 3 months of
the agricultural season).
}
\examples{
\dontrun{
#No examples yet
}
}
\references{
Liess M. & Von der Ohe P. 2005. \emph{Analyzing effects of pesticides on invertebrate
communities in streams}. Environ Toxicol Chem. 24: 954-965.

Wogram J. & Liess M. 2001. \emph{Rank ordering of macroinvertebrate species sensitivity
to toxic compounds by comparison with that of Daphnia magna}. Bull Environ Contam
Toxicol. 67: 360-367

Liess M., Schaefer R., Schriever C. 2008. \emph{The footprint of pesticide stress in
communities - Species traits reveal community effects of toxicants}. Science of
the Total Environment. 406: 484-490
}
\seealso{
\code{\link{StandardiseRawTaxa}}
}
