% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-spear.R
\name{filter_spear}
\alias{filter_spear}
\title{Filter data before calculating SPEAR metric}
\usage{
filter_spear(data, taxa_list = NULL)
}
\arguments{
\item{data}{Dataframe with ecology results...}

\item{taxa_list}{The taxonomic level the sample(s) have been identified at according to
specified taxa lists as described in WFD100 Further Development of River
Invertebrate Classification Tool. Either "TL2" - Taxa List 2, "TL4" - Taxa
List 4 or  "TL5" - Taxa List 5.}
}
\value{
Dataframe of filtered and aggregated results (based on Taxa List) with four
columns: sample_id, label, SPEAR_SPECIES, response
}
\description{
Filter data before calculating SPEAR metric
}
