% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroinvertebrateTaxa.R
\docType{data}
\name{macroinvertebrateTaxa}
\alias{macroinvertebrateTaxa}
\title{Macroinvertebrate Taxa}
\format{
A data frame
}
\usage{
macroinvertebrateTaxa
}
\value{
A data frame with 49 columns and 6656 rows, containing
this key columns:
\itemize{ \item
  \item Taxon name - UNique scientific name of taxon
}
}
\description{
Dataframe of all macroinvertebrate taxa used for WFD100 / RICT
}
\keyword{datasets}
